feat: 完了したワークアウトの表示項目に総重量と合計セット数を追加
